requirements.txt
# Machine Learning
torch
torchvision
scikit-learn

# Quantum Computing
merlinquantum

# Data & Analytics
pandas
numpy
matplotlib
tensorboard

# Utilities
torch-summary
discord-webhook